requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-migrate
psycopg2-binary
twilio    
flask-cors
gunicorn
google-cloud-texttospeech
google-cloud-storage
google-auth